Fixed issue where both ships were executing the same actions
The issue was that both were calling the 'tag to execute' from the scheduler. Instead of using their local tag to execute.

Basically every time the planner is called it generates a new action tag for the agent to execute.
normally this is passed to the specific agent that called it and they store that locally to execute every frame.

What was happening is that every frame all ships were seeing what the latest planner action was generated out of all the agents.