Changed move functions for enemy & added new conditions
Made the move function generic - it's passed a target.

This somewhat changed the behaviour where the enemy ships don't always chase the player, but will still keep shooting. It's a bit strange.

I also added conditions for health checks and far from the health refill station.

This is building up towards having the planner be able to choose to get health