Fixed the no collision issue (spawning still needs fixed)
Had to remove the 'scene default' root node in inside of the 'ship' parent blueprint - this was not present in the original design and is causing issues.

Now the spawning of the enemies does not work - getting 'pending kill' error - which is weird that they're dying on spawn. I used to previously only have this issue when they died.