Added generic function for 'shoot target' to ship, incorporating the check LOS function
There is a slight issue now where the ships will keep trying to execute 'shoot player' but it won't shoot a bullet.

There is no update to the conditions if the ship can't see the player - that needs to be added in.

For when ship can't see the player - to change the condition