Adding shot filtering inside of the projectile code
I tried to do it with collision object types, but it broke the planner for some reason, and I did not have time to debug why. It broke a lot of other things.

Instead I put shot filtering inside of each projectile, it's a bit gross as it's hardcoded for enemy projectiles to all enemy types (this will get messy)