Added energy shield visual around refill stations & made them move
I put the movement on a spline which can be modified in the world outliner.

The energy shield effect took me for ever - the original energy shield effect has a weird reflection when looking straight on top of it like the camera is, which made it impossible to see inside. I still could not figure this out, so I made a 'no reflection' version